Removes non-numbered pagination
